The loader seems to work.
It appears I spent the bulk of my effort on this project in the last month or so writing simple loaders because they were simple, rather than because I expected to need them. I gave little thought to what loaders I'd need.

I suppose it wouldn't be a bad thing to be ready to load things for circumstances yet unforeseen, but loading things for *foreseen* circumstances is even better.